Updated genesis-kit templates
Based on the effort to generalize the genesis kit's Concourse CI pipelines, the genesis-kit templates has been rewritten to facilitate this.  Phase 1 of this rewrite is to provide a generic template that can be used to build a pipeline that has optional support for spec-test job, generation of upstream manifest and releases jobs to import and build new versions if these upstream dependencies change, and to support custom jobs such as bats and cats.